conditional migration for claradmin.. categories_filters is somehow already gone on claradmin (wtf?)

+++ db/migrate/20170809115013_remove_categories_filters_and_categories_section_filters.rb
@@ -1,5 +1,7 @@
 class RemoveCategoriesFiltersAndCategoriesSectionFilters < ActiveRecord::Migration
   def change
-    drop_table :categories_filters
+    if ActiveRecord::Base.connection.table_exists? 'categories_filters'
+      drop_table :categories_filters
+    end
   end
 end
